requirements.txt
certifi==2018.11.29

chardet==3.0.4

defusedxml==0.5.0

dj-places==3.1.0

Django==2.1.4

django-analytical==2.5.0

django-appconf==1.0.2

django-autocomplete-light==3.3.2

django-bootstrap-datepicker-plus==3.0.5

django-crispy-forms==1.7.2

django-select2==6.3.1

django-taggit==0.23.0

django-tinymce4-lite==1.7.3

gunicorn==19.9.0

idna==2.8

jsmin==2.2.2

oauthlib==3.0.0

Pillow==5.4.0

psycopg2==2.7.6.1

pycryptodomex==3.4.7

PyJWT==1.6.4

python3-openid==3.1.0

pytz==2018.7

requests==2.21.0

requests-oauthlib==1.2.0

six==1.12.0

social-auth-app-django==3.1.0

social-auth-core==3.0.0

sorl-thumbnail==12.5.0

stream-django==1.5.0

stream-python==3.0.1

taggit-selectize==2.6.0

urllib3==1.24.1
django-reset-migrations